Crate flux_desugar

Desugaring from types in flux_syntax::surface to types in [flux_middle::fhir]

NOTE

Desugaring requires knowing the sort of each type so we can correctly resolve binders declared with @ syntax or arg syntax. In particular, to know the sort of a type parameter we need to know its kind because only type parameters of sort base can be refined. The essential function implementing this logic is [GlobalEnv::sort_of_res]. This function requires the generics for the item being desugared to be register in [fhir::Map], thus we need to make sure that when desugaring an item, generics are registered before desugaring the rest of the item.

Modules

desugar 🔒

resolver

Statics

DEFAULT_LOCALE_RESOURCE

Raw content of Fluent resource for this crate, generated by fluent_messages macro, imported by rustc_driver to include all crates’ resources in one bundle.

Functions

desugar_defn

desugar_enum_def

desugar_fn_sig

desugar_generics_and_predicates

HACK(nilehmann) this is a bit of a hack. We use it to properly register generics and predicates for items that don’t have surface syntax (impl blocks, traits, …). In this cases we just [lift] them from hir.

desugar_qualifier

desugar_refined_by

desugar_sort_decl

desugar_struct_def

desugar_type_alias

func_def_to_func_decl