flux_desugar
Module desugar

Structs

FluxItemCtxt ğŸ”’

RustItemCtxt ğŸ”’

Traits

DesugarCtxt ğŸ”’

Functions

collect_generics_in_params ğŸ”’

Collect all sorts resolved to a generic type in a list of refinement parameters. Return the set of generic def_ids used (sorted by their position in the list of generics).

desugar_base_sort ğŸ”’

desugar_qualifier ğŸ”’

desugar_sort ğŸ”’

desugar_spec_func ğŸ”’

Type Aliases

Result ğŸ”’