Struct flux_desugar::desugar::env::Env

pub(crate) struct Env<P> {
    scopes: FxHashMap<ScopeId, Scope<P>>,
    parent: FxHashMap<ScopeId, ScopeId>,
    children: FxHashMap<ScopeId, FxHashSet<ScopeId>>,
    root: ScopeId,
    curr: ScopeId,
}

The environment used for desugaring that tracks parameter names and shadowing. An environment is a tree formed by nested scopes.

Fields
scopes: FxHashMap<ScopeId, Scope<P>>parent: FxHashMap<ScopeId, ScopeId>children: FxHashMap<ScopeId, FxHashSet<ScopeId>>root: ScopeIdcurr: ScopeId
Implementations

impl<P> Env<P>

pub(crate) fn new(root: ScopeId) -> Self

pub(crate) fn insert( &mut self, sess: &FluxSession, ident: Ident, param: P ) -> Result<(), ErrorGuaranteed>

Inserts a parameter in the current scope, returning an error if a parameter with the same name is already in the scope.

pub(crate) fn extend( &mut self, sess: &FluxSession, params: impl IntoIterator<Item = (Ident, P)> ) -> Result<(), ErrorGuaranteed>

Extends the current scope with a list of parameters reporting an error if there are any dupliates.

pub(crate) fn get(&self, ident: Ident) -> Option<&P>

pub(crate) fn get_with_scope(&self, ident: Ident) -> Option<(ScopeId, &P)>

fn find_map<'a, T>( &'a self, f: impl FnMut(ScopeId, &'a Scope<P>) -> Option<T> ) -> Option<T>

pub(crate) fn get_mut(&mut self, ident: Ident) -> Option<&mut P>

pub(crate) fn scope(&mut self, id: ScopeId) -> &mut Scope<P>

pub(crate) fn push(&mut self, id: ScopeId)

Push a scope with id as a child of the current scope and then enters the new scope. The id must be different from any scope already in the environment.

Panics

Panics if id is already in the enrvionment

pub(crate) fn pop(&mut self) -> Scope<P>

Remove the current scope and return it. Then set the current scope to the parent scope.

Panics

Panics if the current scope has children.

pub(crate) fn enter(&mut self, scope: ScopeId)

Enter a child scope.

Panics

Panics if scope is not a child of the current scope.

pub(crate) fn exit(&mut self)

Exit the current scope back into the parent scope.

Panics

Panics if the current scope is the root scope.

pub(crate) fn filter_map<T>(self, f: impl FnMut(P, bool) -> Option<T>) -> Env<T>

fn curr(&mut self) -> &mut Scope<P>

pub(crate) fn root(&mut self) -> &mut Scope<P>

pub(crate) fn into_root(self) -> Scope<P>

impl<P: Debug> Env<P>

fn fmt_rec( &self, scope_id: ScopeId, prefix: &str, is_last: bool, f: &mut Formatter<'_> ) -> Result

impl Env<Param>

fn into_desugar_env(self) -> Env<Param>

impl Env<Param>

fn from_params<'a>( genv: &GlobalEnv<'_, '_>, sort_resolver: &SortResolver<'_>, scope: ScopeId, params: impl IntoIterator<Item = &'a RefineParam> ) -> Result<Self, ErrorGuaranteed>