flux_desugar::desugar
Function collect_generics_in_params

fn collect_generics_in_params(
    genv: GlobalEnv<'_, '_>,
    owner: MaybeExternId<OwnerId>,
    resolver_output: &ResolverOutput,
    params: &RefineParams,
) -> FxIndexSet<DefId>

Collect all sorts resolved to a generic type in a list of refinement parameters. Return the set of generic def_ids used (sorted by their position in the list of generics).