Struct flux_desugar::desugar::Binders

pub(crate) struct Binders {
    name_gen: IndexGen<Name>,
    layers: Vec<Layer>,
}

Keeps track of the surface level identifiers in scope and a mapping between them and a Binder.

Fields
name_gen: IndexGen<Name>layers: Vec<Layer>
Implementations

impl Binders

pub(crate) fn new() -> Binders

fn from_params<'a>( genv: &GlobalEnv<'_, '_>, params: impl IntoIterator<Item = &'a RefineParam> ) -> Result<Self, ErrorGuaranteed>

fn insert_params<'a>( &mut self, genv: &GlobalEnv<'_, '_>, params: impl IntoIterator<Item = &'a RefineParam> ) -> Result<(), ErrorGuaranteed>

fn fresh(&self) -> Name

fn get(&self, ident: impl Borrow<Ident>) -> Option<&Binder>

fn insert_binder( &mut self, sess: &FluxSession, ident: Ident, binder: Binder ) -> Result<(), ErrorGuaranteed>

fn bot_layer(&mut self) -> &mut Layer

fn top_layer(&mut self) -> &mut Layer

fn iter_layers<'a, T>( &'a self, f: impl FnMut(&'a Layer) -> Option<T> ) -> Option<T>

fn push_layer(&mut self)

fn pop_layer(&mut self) -> Layer

fn binder_from_sort(&self, sort: Sort) -> Binder

fn binder_from_res(&self, genv: &GlobalEnv<'_, '_>, res: Res) -> Binder

fn binder_from_bty( &self, genv: &GlobalEnv<'_, '_>, resolver_output: &ResolverOutput, bty: &BaseTy ) -> Binder