Module flux_desugar::desugar

Modules

errors ğŸ”’

Macros

define_resolve_num_const ğŸ”’

Structs

Binders ğŸ”’

Keeps track of the surface level identifiers in scope and a mapping between them and a Binder.

DesugarCtxt ğŸ”’

ExprCtxt ğŸ”’

Layer ğŸ”’

Sorts ğŸ”’

Enums

Binder ğŸ”’

The different kind of binders that can appear in the surface syntax

FuncRes ğŸ”’

QPathRes ğŸ”’

TypePos ğŸ”’

Statics

SORTS ğŸ”’

Traits

DesugarContext ğŸ”’

Functions

as_tuple ğŸ”’

desugar_bin_op ğŸ”’

desugar_defn

desugar_qualifier

desugar_refined_by

desugar_un_op ğŸ”’

func_def_to_func_decl

index_sort ğŸ”’

resolve_app_sort ğŸ”’

resolve_base_sort ğŸ”’

resolve_base_sort_ident ğŸ”’

resolve_func_sort ğŸ”’

resolve_num_const ğŸ”’

resolve_sort ğŸ”’

resolve_sort_ctor ğŸ”’