Struct flux_desugar::desugar::RustItemCtxt

pub(crate) struct RustItemCtxt<'a, 'tcx> {
    genv: &'a GlobalEnv<'a, 'tcx>,
    local_id_gen: IndexGen<ItemLocalId>,
    owner: OwnerId,
    resolver_output: &'a ResolverOutput,
    opaque_tys: Option<&'a mut UnordMap<LocalDefId, OpaqueTy>>,
    sort_resolver: SortResolver<'a>,
}

Fields
genv: &'a GlobalEnv<'a, 'tcx>local_id_gen: IndexGen<ItemLocalId>owner: OwnerIdresolver_output: &'a ResolverOutputopaque_tys: Option<&'a mut UnordMap<LocalDefId, OpaqueTy>>sort_resolver: SortResolver<'a>
Implementations

impl RustItemCtxt<'_, '_>

pub(super) fn gather_params_type_alias( &self, ty_alias: &TyAlias ) -> Result<Env<Param>, ErrorGuaranteed>

pub(super) fn gather_params_struct( &self, struct_def: &StructDef ) -> Result<Env<Param>, ErrorGuaranteed>

pub(super) fn gather_params_variant( &self, variant_def: &VariantDef ) -> Result<Env<Param>, ErrorGuaranteed>

fn gather_params_variant_ret( &self, ret: &VariantRet, env: &mut Env<Param> ) -> Result<(), ErrorGuaranteed>

pub(super) fn gather_params_fn_sig( &mut self, fn_sig: &FnSig ) -> Result<Env<Param>, ErrorGuaranteed>

fn gather_params_fn_sig_input( &self, fn_sig: &FnSig, env: &mut Env<Param> ) -> Result<(), ErrorGuaranteed>

fn gather_params_predicates( &self, predicates: &[WhereBoundPredicate], env: &mut Env<Param> ) -> Result<(), ErrorGuaranteed>

Synthetic parameters cannot be defined inside predicates but we traverse it to report errors if we find them.

fn gather_params_fn_sig_output( &self, fn_sig: &FnSig, env: &mut Env<Param> ) -> Result<(), ErrorGuaranteed>

fn gather_params_fun_arg( &self, arg: &Arg, env: &mut Env<Param> ) -> Result<(), ErrorGuaranteed>

fn gather_params_ty( &self, bind: Option<Ident>, ty: &Ty, pos: TypePos, env: &mut Env<Param> ) -> Result<(), ErrorGuaranteed>

fn gather_params_indices( &self, indices: &Indices, pos: TypePos, env: &mut Env<Param> ) -> Result<(), ErrorGuaranteed>

fn gather_params_refine_arg( &self, arg: &RefineArg, pos: TypePos, env: &mut Env<Param> ) -> Result<(), ErrorGuaranteed>

fn gather_params_path( &self, path: &Path, pos: TypePos, params: &mut Env<Param> ) -> Result<(), ErrorGuaranteed>

fn gather_params_generic_arg( &self, arg: &GenericArg, pos: TypePos, params: &mut Env<Param> ) -> Result<(), ErrorGuaranteed>

fn gather_params_bty( &self, bty: &BaseTy, pos: TypePos, params: &mut Env<Param> ) -> Result<(), ErrorGuaranteed>

fn check_param_uses( &self, env: &mut Env<Param>, f: impl FnOnce(&mut CheckParamUses<'_>) ) -> Result<(), ErrorGuaranteed>

fn resolve_params<'a>( &self, params: impl IntoIterator<Item = &'a RefineParam> ) -> Result<Vec<(Ident, Param)>, ErrorGuaranteed>

impl<'a, 'tcx> RustItemCtxt<'a, 'tcx>

pub(crate) fn new( genv: &'a GlobalEnv<'a, 'tcx>, owner: OwnerId, resolver_output: &'a ResolverOutput, opaque_tys: Option<&'a mut UnordMap<LocalDefId, OpaqueTy>> ) -> RustItemCtxt<'a, 'tcx>

fn with_new_owner<'b>(&'b mut self, owner: OwnerId) -> RustItemCtxt<'b, 'tcx>

pub(crate) fn as_lift_cx<'b>(&'b mut self) -> LiftCtxt<'b, 'tcx>

pub(crate) fn desugar_generics( &self, generics: &Generics ) -> Result<Generics, ErrorGuaranteed>

fn desugar_predicates( &mut self, predicates: &Vec<WhereBoundPredicate>, env: &mut Env<Param> ) -> Result<GenericPredicates, ErrorGuaranteed>

fn desugar_predicate( &mut self, pred: &WhereBoundPredicate, env: &mut Env<Param> ) -> Result<WhereBoundPredicate, ErrorGuaranteed>

fn desugar_generic_bounds( &mut self, bounds: &GenericBounds, env: &mut Env<Param> ) -> Result<GenericBounds, ErrorGuaranteed>

fn desugar_trait_ref( &mut self, trait_ref: &TraitRef, env: &mut Env<Param> ) -> Result<PolyTraitRef, ErrorGuaranteed>

pub(crate) fn desugar_struct_def( &mut self, struct_def: &StructDef ) -> Result<StructDef, ErrorGuaranteed>

pub(crate) fn desugar_enum_def( &mut self, enum_def: &EnumDef ) -> Result<EnumDef, ErrorGuaranteed>

fn desugar_enum_variant_def( &mut self, variant_def: &Option<VariantDef>, hir_variant: &Variant<'_> ) -> Result<VariantDef, ErrorGuaranteed>

pub(crate) fn desugar_type_alias( &mut self, ty_alias: &TyAlias ) -> Result<TyAlias, ErrorGuaranteed>

pub(crate) fn desugar_fn_sig( &mut self, fn_sig: &FnSig ) -> Result<(GenericPredicates, FnSig), ErrorGuaranteed>

fn desugar_constraint( &mut self, cstr: &Constraint, env: &mut Env<Param> ) -> Result<Constraint, ErrorGuaranteed>

fn desugar_fun_arg( &mut self, arg: &Arg, env: &mut Env<Param>, requires: &mut Vec<Constraint> ) -> Result<Ty, ErrorGuaranteed>

fn desugar_asyncness( &mut self, asyncness: Async, returns: &FnRetTy, env: &mut Env<Param> ) -> Result<Ty, ErrorGuaranteed>

fn desugar_opaque_ty_for_async( &mut self, returns: &FnRetTy, env: &mut Env<Param> ) -> Result<OpaqueTy, ErrorGuaranteed>

fn desugar_fn_ret_ty( &mut self, returns: &FnRetTy, env: &mut Env<Param> ) -> Result<Ty, ErrorGuaranteed>

fn desugar_ty( &mut self, bind: Option<Ident>, ty: &Ty, env: &mut Env<Param> ) -> Result<Ty, ErrorGuaranteed>

fn desugar_opaque_ty_for_impl_trait( &mut self, bounds: &GenericBounds, env: &mut Env<Param> ) -> Result<OpaqueTy, ErrorGuaranteed>

fn mk_lft_hole(&self) -> Lifetime

fn desugar_indices( &mut self, bty: &BaseTy, idxs: &Indices, env: &mut Env<Param> ) -> Result<RefineArg, ErrorGuaranteed>

fn desugar_refine_arg( &mut self, arg: &RefineArg, sort: Option<Sort>, env: &mut Env<Param> ) -> Result<RefineArg, ErrorGuaranteed>

fn bind_into_refine_arg( &self, ident: Ident, sort: Option<Sort>, env: &mut Env<Param> ) -> Result<Option<RefineArg>, ErrorGuaranteed>

fn desugar_bty( &mut self, bty: &BaseTy, env: &mut Env<Param> ) -> Result<BaseTy, ErrorGuaranteed>

fn desugar_path( &mut self, path: &Path, env: &mut Env<Param> ) -> Result<Path, ErrorGuaranteed>

fn desugar_path_to_bty( &mut self, path: &Path, env: &mut Env<Param> ) -> Result<BaseTy, ErrorGuaranteed>

fn desugar_generic_args( &mut self, res: Res, args: &[GenericArg], env: &mut Env<Param> ) -> Result<(Vec<GenericArg>, Vec<TypeBinding>), ErrorGuaranteed>

fn desugar_bty_bind( &mut self, bind: Option<Ident>, bty: &BaseTy, env: &mut Env<Param> ) -> Result<Ty, ErrorGuaranteed>

fn desugar_variant_ret( &mut self, ret: &VariantRet, env: &mut Env<Param> ) -> Result<VariantRet, ErrorGuaranteed>

fn insert_opaque_ty(&mut self, def_id: LocalDefId, opaque_ty: OpaqueTy)

fn sess(&self) -> &'a FluxSession

fn emit_err<'b>(&'b self, err: impl IntoDiagnostic<'b>) -> ErrorGuaranteed